requirements.txt
python-telegram-bot==20.3
requests==2.31.0
motor==3.2.0
pytz==2023.3
